app/requirements.txt
argcomplete==3.5.0
asgiref==3.8.1
boto3==1.35.21
botocore==1.35.21
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-filter==24.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-yasg==1.21.7
durationpy==0.7
gunicorn==23.0.0
hjson==3.1.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kappa==0.6.0
mangum==0.17.0
MarkupSafe==2.1.5
packaging==24.1
placebo==0.9.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
whitenoise==6.7.0
zappa==0.59.0
